fix: split turn integration tests under 400-line CI cap (#3876)
Replace collapsed integration entrypoints with standalone discoverable
_test.dart files (no part-of) and extract world-market trade scenario
fixtures so repo.domain_package_test_file_size passes.

Refs #3876
